Move tagging implementation classes into io.opencensus.impl.tags. (part of #379)
This change will ensure that the 'tags' implementation only accesses the API
classes through their public API.  That will help us maintain backwards
compatibility.
