+++ core/src/main/java/io/opencensus/tags/Tags.java
@@ -42,7 +42,7 @@ public final class Tags {
     try {
       // Call Class.forName with literal string name of the class to help shading tools.
       return Provider.createInstance(
-          Class.forName("io.opencensus.tags.TagsComponentImpl", true, classLoader),
+          Class.forName("io.opencensus.impl.tags.TagsComponentImpl", true, classLoader),
           TagsComponent.class);
     } catch (ClassNotFoundException e) {
       logger.log(
@@ -54,7 +54,7 @@ public final class Tags {
     try {
       // Call Class.forName with literal string name of the class to help shading tools.
       return Provider.createInstance(
-          Class.forName("io.opencensus.tags.TagsComponentImplLite", true, classLoader),
+          Class.forName("io.opencensus.impl.tags.TagsComponentImplLite", true, classLoader),
           TagsComponent.class);
     } catch (ClassNotFoundException e) {
       logger.log(
