+++ core_impl_java/src/test/java/io/opencensus/impl/tags/TagsTest.java
@@ -11,10 +11,12 @@
  * limitations under the License.
  */
 
-package io.opencensus.tags;
+package io.opencensus.impl.tags;
 
 import static com.google.common.truth.Truth.assertThat;
 
+import io.opencensus.tags.Tags;
+import io.opencensus.tags.TagsComponent;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
